requirements.txt
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.0
streamlit>=1.29.0
plotly>=5.18.0
python-dotenv>=1.0.0
scipy>=1.11.0
boto3>=1.40.0
pytz>=2024.1
